Move s3_recipe_helper in to common directory to reuse in other tests. (#6535)
