+++ ydb/library/testlib/s3_recipe_helper/ya.make
@@ -1,0 +1,11 @@
+LIBRARY()
+
+SRCS(
+    s3_recipe_helper.cpp
+)
+
+PEERDIR(
+    contrib/libs/aws-sdk-cpp/aws-cpp-sdk-s3
+)
+
+END()
